requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.6.0+cu118
torch_geometric==2.6.1
ogb==1.3.6
PyYAML==6.0.1
matplotlib==3.3.4